Renamed manager to tracker
